requirements.txt
# core
batchgenerators==0.21
dcm2niix
fury==0.9.0
nibabel==5.1.0
nnunet-customized==1.2
numpy==1.23.5
p_tqdm==1.4.0
psutil==5.9.5
pydantic==1.10.9
pyradiomics==3.0.1
python-dotenv==1.0.0
requests==2.31.0
rt_utils==1.2.7
SimpleITK==2.2.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.48.2
typer==0.9.0
types-requests==2.31.0.1
xvfbwrapper==0.2.9
xmltodict==0.13.0
pyyaml==6.0
wandb==0.15.4
# gpu
cucim==23.4.1
cupy-cuda11x==12.1.0
# curadiomics
tensorflow==2.12.0
# dev
pre-commit
# test
pytest
